build-system: update android/build.sh to use get-dep-lib.sh
Remark: the openssl version is used in a cmake call.

android/build.sh uses wget and curl is not present, so added a
switch in get-dep-lib to choose between curl and wget.

Adding the curl command to the wrapper script does not seem like
an option, because this script is supposed to be used by developers
on different platforms.
(e.g. it tests for the existence of commands but do not add them)

Android also uses old versions of libzip, libusb and openssl,
never version does not compile.

For those 2 reasons an exception is added to get-dep-lib for now,
awaiting an android developer

Signed-off-by: Jan Iversen <jani@apache.org>