Travis: add secondary Linux build target
The goal is to reduce the overall wall clock time of our test builds.

This secondary Linux target uses the no-sudo container setup with a Qt
5.10.1 backport and tests the build without BT support as well as the
mobile build (but not the full desktop build and doesn't create an
AppImage as the pre-built Qt5.10.1 is missing QtWebKit).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>

Travis: use Qt 5.10.1 packages for Linux

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>