Bluetooth: simulate Linux/Mac behavior on Windows
Qt's Linux/Mac implementations of the BT scanning code are asynchronous
and can be started/stopped at any time. In contrast, our own Windows
implementation, even though it runs in its own thread, is in principle
synchronous, because it hangs in syscalls. On start/stop, etc we wait
for the thread and accordingly the syscall to finish.

Simulate the Linux/Mac behavior by implementing an event-loop that
runs as long as a running-flag is set. In a way this is cheating -
the results will only appear in batches, but it helps unify the code
until Qt implements their own Windows version.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>