build-system: add travis_wait shell functions
These come originally from https://github.com/travis-ci/travis-build and are
available when running on Travis, but not when running inside a Docker
container on Travis as we do in order to build Android.

The goal is to provide a quasi heart-beat on STDOUT during very long running
commands - for us the wget to download Qt often times out, so that's where we
are going to use this.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>