Bluetooth: adapt memory management of BluetoothDeviceDiscoveryAgent
On Linux/Mac, replace the raw-pointers of localDevice and
remoteDeviceDiscoveryAgent by QScopedPointers. This simplifies
memory-management, as these objects need not be manually deleted
anymore. Indeed, this fixes a leak, whereby the removeDeviceDiscoveryAgent
would leak if the local device was changed.

On Windows, don't allocate the remoteDeviceDiscoveryAgent, but
use a global object. This is necessary, because we don't want
delays if the agent is stuck in a syscall.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>