Bluetooth: don't assume a length SOCKADDR structure on Windows
The old code assumes that WSALookupServiceNextW() returns
a SOCKADDR of type SOCKADDR_BTH and passes the according length
to WSAAddressToStringW.

Since we're not using the content of this structure directly,
simply pass the length returned in WSALookupServiceNextW()
to WSAAddressToStringW(). This makes the code cleaner in
that it avoids a SOCKADDR->SOCKADDR_BTH->SOCKADDR down-and-up
cast.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>