Tests: work around odd compile failure on Travis
This seems silly, but including btdiscovery.h in testhelper.cpp fails on
Travis with an include that can't be found - even though that same
include is found elsewhere in the build process.

Instead of trying to fight that, simply declare the two functions that
we are testing and be done with it.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>